Separate stopword filter from hasher
